% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z11_download_data.R
\name{z11_download_data}
\alias{z11_download_data}
\title{Download Census 2011 data}
\usage{
z11_get_data(name, directory, all = FALSE)
}
\arguments{
\item{directory}{Path to the directory where the data should be saved.}
}
\description{
Downloads Census 2011 data, unzips it and saves it to a local folder
}
\details{
This was written on Ubuntu 18, and I can't guarantee that it works on anything else. If you're on Windows or Mac, download and unzip the data manually instead.
}
\examples{
z11_download_data("/home/yourname/z11data")

}
