% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z11_download_links.R
\docType{data}
\name{z11_download_links}
\alias{z11_download_links}
\title{Download Links for Zensus 2011}
\format{
A named character vector
}
\source{
\href{https://www.zensus2011.de/DE/Home/home_node.html}{Zensus 2011}
}
\usage{
data(z11_download_links)
}
\description{
This is where the Zensus 2011 .csv files can be downloaded. Is used by the z11_get_data() function.
}
\examples{
data(z11_download_links)
z11_download_links["bevoelkerung100m"]
}
\keyword{datasets}
