% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z11_extract_inspire_coordinates.R
\name{z11_extract_inspire_coordinates}
\alias{z11_extract_inspire_coordinates}
\title{Extract Geocoordinates (EPSG:3035) from INSPIRE IDs}
\usage{
z11_extract_inspire_coordinates(inspire_ids)
}
\arguments{
\item{inspire_ids}{Character vector containing INSPIRE IDs}
}
\value{
tibble
}
\description{
Retreive centroid geocoordinates for 1km^2 and 1 hectare INSPIRE Grid IDs
}
