% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z11_get_100m_attribute.R
\name{z11_get_100m_attribute}
\alias{z11_get_100m_attribute}
\alias{z11_get_100m_attribute.DBIConnection}
\alias{z11_get_100m_attribute.default}
\title{Retrieve 1 hectare attribute of the German Census 2011}
\usage{
z11_get_100m_attribute(
  attribute,
  data_source,
  geometry = TRUE,
  as_raster = TRUE
)

\method{z11_get_100m_attribute}{DBIConnection}(
  attribute,
  data_source,
  geometry = TRUE,
  as_raster = TRUE
)

\method{z11_get_100m_attribute}{default}(
  attribute,
  data_source,
  geometry = TRUE,
  as_raster = TRUE
)
}
\arguments{
\item{attribute}{attribute name as definded in
\code{z11::z11_list_100m_attributes}}

\item{data_source}{Either a DBI connection or a character string containing
a file path to the downloaded census data https://github.com/StefanJuenger/z11data;
default is NULL - data are downloaded from the internet}

\item{geometry}{logical; should coordinates be extracted from the INSPIRE ID?}

\item{as_raster}{logical; shall the attribute be returned as raster or sf
object}
}
\value{
Raster or sf
}
\description{
This function retrieves an attribute with a raster resolution of 1 hectare.
You can either choose to have it converted directly into the raster format
(default) or you can return it as a `sf` tibble object with point
geometries.
}
