% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z11_get_1km_attribute_db.R
\name{z11_get_1km_attribute_db}
\alias{z11_get_1km_attribute_db}
\title{Retrieve 1km^2 attribute of the German Census 2011}
\usage{
z11_get_1km_attribute_db(attribute, con, as_raster = TRUE)
}
\arguments{
\item{attribute}{attribute name as defined in
\code{z11::z11_list_1km_attributes}, as a string}

\item{con}{Connection to the database}

\item{as_raster}{logical; shall the attribute be returned as raster or sf
object}
}
\value{
Raster or sf
}
\description{
This function retrieves an attribute with a raster resolution of 1km². You
can either choose to have it converted directly into the raster format
(default) or you can return it as a \code{sf} tibble object with point
geometries.
}
