% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z11_get_data.R
\name{z11_get_data}
\alias{z11_get_data}
\title{Download Zensus 2011 data}
\usage{
z11_get_data(name, directory, all = FALSE)
}
\arguments{
\item{name}{Which dataset should be downloaded? Possible values are "klassiert1km", "spitz1km", "bevoelkerung100m", "demographie100m", "familien100m", "haushalte100m", "gebaude100m", "wohnungen100m".}

\item{directory}{Path to the directory where the data should be saved.}

\item{all}{Should all datasets be downloaded? Defaults to FALSE. Use with caution.}
}
\description{
Downloads Zensus 2011 data, splits it into smaller chunks, and saves it locally.
}
\details{
The data files are very large and use a lot of RAM, so use with caution. Don't run other things in the background.
This was tested on Ubuntu 18, and I can't guarantee that it works on anything else.
}
\examples{
z11_get_data("spitz1km", "/home/yourname/z11data")

}
