% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z11_join_db_100m.R
\name{z11_join_db_100m}
\alias{z11_join_db_100m}
\title{Join 100m Census attributes}
\usage{
z11_join_db_100m(df, inspire_column, con, var = NULL)
}
\arguments{
\item{df}{A dataframe}

\item{con}{Connection to a database. Open with DBI::dbConnect, and remember to close it later.}

\item{var}{The name of a Census variable. Optional. If no variable name is given, all 100m attributes will be joined to the dataframe.}

\item{A}{string containing the name of the 100m INSPIRE ID in the dataframe}
}
\description{
Takes a dataframe as input, and joins 100m Census attributes on the INSPIRE ID
}
\examples{
con <- dbConnect(RSQLite::SQLite(), "/path/to/db.sqlite3")
z11_join_db_100m(df, "inspire_id", con, "GEB_HEIZTYP_1")
dbDisconnect(con)

}
