% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z11_join_db_1km.R
\name{z11_join_db_1km}
\alias{z11_join_db_1km}
\title{Join 1km Census attributes}
\usage{
z11_join_db_1km(df, inspire_column, con, var = NULL)
}
\arguments{
\item{df}{A dataframe}

\item{con}{Connection to a database. Open with DBI::dbConnect, and remember to close it later.}

\item{var}{The name of a Census variable. Optional. If no variable name is given, all 1km attributes will be joined to the dataframe.}

\item{A}{string containing the name of the 1km INSPIRE ID in the dataframe}
}
\description{
Takes a dataframe as input, and joins 1km Census attributes on the INSPIRE ID
}
\examples{
con <- dbConnect(RSQLite::SQLite(), "/path/to/db.sqlite3")
z11_join_db_1km(df, "inspire_id", con, "Frauen_A_cat")
dbDisconnect(con)

}
