% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z11_list_100m_attributes_db.R
\name{z11_list_100m_attributes_db}
\alias{z11_list_100m_attributes_db}
\title{List 1 hectare attributes of the German Census 2011}
\usage{
z11_list_100m_attributes_db(con)
}
\value{
List
}
\description{
This function lists all attributes of the German Census 2011 with a raster
resolution of 1 hectare. These data were the last that have been published.
While they comprise more attributes than the 1 km² ones, the data were
prepared in a less easy to be used format. Note that the returned vector
of attribute names includes the attribute names as they were defined by the
Census folks plus a pattern denoting the code of the attribute in the
original dumped CSV file.
}
