% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z11_list_1km_attributes.R
\name{z11_list_1km_attributes}
\alias{z11_list_1km_attributes}
\title{List 1km^2 attributes of the German Census 2011}
\usage{
z11_list_1km_attributes()
}
\value{
Character vector
}
\description{
This function lists all attributes of the German Census 2011 with a raster
resolution of 1km^2. These data were the first that have been published. And
while they comprise fewer attributes than the 1 hectare ones, the data were
prepared in a way more easy to be used format. Note that the returned vector
of attribute names includes the attribute names as they were defined by the
Census folks.
}
