% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z11_prepare_data.R
\name{z11_prepare_data}
\alias{z11_prepare_data}
\title{Data Preparation}
\usage{
z11_prepare_data(file, data_location)
}
\description{
Reads in 2011 Census .csv files, splits it into smaller chunks, and saves it
locally.
}
