% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z11_simple_join_100m_attribute.R
\name{z11_simple_join_100m_attribute}
\alias{z11_simple_join_100m_attribute}
\title{Simple join of data with census 100m attributes}
\usage{
z11_simple_join_100m_attribute(
  df,
  inspire_column,
  data_source = NULL,
  attributes = NULL
)
}
\arguments{
\item{data}{input data}

\item{inspire_colum}{Character string for column name in input data
containing the inspire ID}

\item{attribute}{Name of the census variable which is aimed to be merged with
input data (used for calling \code{z11::z11_get_100m_attribute()})}

\item{all}{logical; should all census attributes be merged? default: `FALSE`}

\item{...}{arguments passed to \code{z11::z11_get_100m_attribute()}}
}
\description{
Merge input data with census 100m data by simple matching of INSPIRE IDs
}
