% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z11_wide.R
\name{z11_wide_100m}
\alias{z11_wide_100m}
\title{Convert German Census 100m X 100m long format files in wide format}
\usage{
z11_wide_100m(x)
}
\arguments{
\item{x}{R object containing the census data}
}
\description{
This function expects an R object containing the census data in the long
format. It processes the data as following: \\
1. Deleting all non-substantial rows from the data (e.g. "INSGESAMT"
rows) \\
2. Creating identifier for all attribute-code-combinations \\
3. Transforming of data using the previously created identifier \\
4. Lastly, all missing values are converted to zeros
}
