% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z11_write_to_db.R
\name{z11_write_to_db}
\alias{z11_write_to_db}
\title{Add 2011 Census data to SQLite Database}
\usage{
z11_write_to_db(name, file = NULL, con)
}
\arguments{
\item{file}{Path to the .csv file}

\item{con}{Connection to a database. Open with DBI::dbConnect, and remember to close it later.}
}
\description{
Add a Census .csv file to a SQLite Database
}
\details{
The data files are very large and use a lot of RAM, so use with caution. Don't run other things in the background.
}
\examples{
con <- dbConnect(RSQLite::SQLite(), "/path/to/db.sqlite3")
z11_fill_db("/home/yourname/Familien100m.csv", con = con)
dbDisconnect(con)

}
