% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z11_get_100m_attribute.R
\name{z11_get_100m_attribute}
\alias{z11_get_100m_attribute}
\title{Retrieve 1 hectare attribute of the German Census 2011}
\usage{
z11_get_100m_attribute(attribute, as_raster = TRUE, data_location = NULL)
}
\arguments{
\item{attribute}{attribute name as definded in
\code{z11::z11_list_100m_attributes}}

\item{as_raster}{logical; shall the attribute be returned as raster or sf
object}

\item{data_location}{character string; location of the downloaded census data
from https://github.com/StefanJuenger/z11data; default is NULL - data are
downloaded from the internet}
}
\value{
Raster or sf
}
\description{
This function retrieves an attribute with a raster resolution of 1 hectare.
You can either choose to have it converted directly into the raster format
(default) or you can return it as a `sf` tibble object with point
geometries.
}
