% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z11_prepare_data.R
\name{z11_prepare_data}
\alias{z11_prepare_data}
\title{Data Preparation}
\usage{
z11_prepare_data(file, directory)
}
\arguments{
\item{file}{Path to a either a census .csv file or a directory where the .csv files are saved}

\item{directory}{Path to the directory where the data should be saved.}
}
\description{
Reads in 2011 Census .csv files, splits it into smaller chunks, and saves it locally.
}
\details{
The data files are very large and use a lot of RAM, so use with caution. Don't run other things in the background.
}
\examples{
z11_prepare_data("/home/yourname/Haushalte100m.csv", "/home/yourname/z11data")

}
