% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z11_simple_join_100m_attribute.R
\name{z11_simple_join_100m_attribute}
\alias{z11_simple_join_100m_attribute}
\alias{z11_simple_join_100m_attribute.DBIConnection}
\alias{z11_simple_join_100m_attribute.default}
\title{Simple join of data with census 100m attributes}
\usage{
z11_simple_join_100m_attribute(
  df,
  inspire_column,
  data_source = NULL,
  attributes = NULL
)

\method{z11_simple_join_100m_attribute}{DBIConnection}(
  df,
  inspire_column,
  data_source,
  attributes = NULL
)

\method{z11_simple_join_100m_attribute}{default}(
  df,
  inspire_column,
  data_source,
  attributes = NULL
)
}
\arguments{
\item{df}{input data}

\item{inspire_column}{Column name in input data containing the inspire ID}

\item{data_source}{Either a DBI connection, or a character string containing
a file path to the data location}

\item{attributes}{A character or character vector containing the name of the
Census attribute to be merged with the input data. If no attribute name is
given, all available attributes will be merged.}
}
\description{
Merge input data with census 100m data by simple matching of INSPIRE IDs
}
\examples{
\dontrun{joined <- z11_simple_join_100m_attribute(df, inspire_100m, con)}

}
